requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.3.2
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.1
h11==0.14.0
html2text==2020.1.16
identify==2.5.35
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
MouseInfo==0.1.3
multidict==6.0.4
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
requests==2.31.0
responses==0.24.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
twilio==8.11.0
tzlocal==5.2
urllib3==2.1.0
virtualenv==20.25.1
webdriver-manager==4.0.1
Werkzeug==3.0.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.4
